requirements.txt
Django==5.0.6
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.5
django-jazzmin==3.0.0
django-modeltranslation==0.19.6
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
ipython==8.26.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==72.1.0